Improve appEnv 'visibility' for functions running in WebbyM
With this change appEnv is only an 'ask' away. WebbyM handlers can work
with functions of the form,

  `doSomething :: (HasSomeConfig env, MonadReader env m) -> m a`

without additional boiler-plate.

Previously, since appEnv was tucked in alongside WebbyM's internal state, held
as a ReaderT environment it wasn't available through the MonadReader's 'ask'
method. This meant, for every 'appEnv', 'HasSomeConfig appEnv' combination the
library user had to define a typeclass of the following form,

  `class MonadSomeConfig m where
      getSomeConfigM :: m SomeConfig
  `

and implement the instance for 'WebbyM appEnv' which would simply fetch it from
the appEnv.